remove pages compare after execution (#1275)

+++ runtime-interface/src/deprecated.rs
@@ -104,7 +104,7 @@ pub(crate) fn mprotect_pages_slice(
     let mprotect = |start: PageNumber, count, protect: bool| unsafe {
         let addr = mem_addr + (start.0 as usize * PageNumber::size()) as HostPointer;
         let size = count as usize * PageNumber::size();
-        sys_mprotect_interval(addr, size, !protect, !protect, false)
+        sys_mprotect_interval(addr as usize, size, !protect, !protect, false)
     };
 
     // Collects continuous intervals of memory from lazy pages to protect them.
